# odd-platform-specification/openapi.yaml
openapi: 3.0.3
info:
  title: ProspectLog data catalog HTTP API contract
  description: ProspectLog data catalog HTTP API contract
  contact:
    email: ndementev@provectus.com
    url: https://provectus.com
    name: Nikita Dementev
  version: 0.0.1
servers:
  - url: 'http://localhost'
    description: stub
...
paths:
...
  /api/datasources:
    get:
      summary: List of data sources
      description: Gets the list of available data sources
      operationId: getDataSourceList
      parameters:
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
        - $ref: './components.yaml/#/components/parameters/SearchParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataSourceList'
      tags:
        - dataSource
    post:
      summary: Register a data source
      description: Registers a data source
      operationId: registerDataSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/DataSourceFormData'
      responses:
        '201':
          description: The resource has been successfully created
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataSource'
      tags:
        - dataSource
  /api/datasources/{data_source_id}:
    put:
      summary: Update data source
      description: Updates data source
      operationId: updateDataSource
      parameters:
        - name: data_source_id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/DataSourceUpdateFormData'
      responses:
        '201':
          description: The resource has been successfully updated
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataSource'
      tags:
        - dataSource
    delete:
      summary: Delete data source
      description: Deletes data source
      operationId: deleteDataSource
      parameters:
        - name: data_source_id
          in: path
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '204':
          $ref: './components.yaml/#/components/responses/Deleted'
      tags:
        - dataSource
...
  /api/dataentities/my:
    get:
      summary: Get my objects
      description: Returns list of data entities owned by current user
      operationId: getMyObjects
      parameters:
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataEntityRefList'
      tags:
        - dataEntity
  /api/dataentities/my/upstream:
    get:
      summary: Get my objects with upstream dependencies
      description: Returns list of data entities owned by current user with upstream dependencies
      operationId: getMyObjectsWithUpstream
      parameters:
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataEntityRefList'
      tags:
        - dataEntity
  /api/dataentities/my/downstream:
    get:
      summary: Get my objects with downstream dependencies
      description: Returns list of data entities owned by current user with downstream dependencies
      operationId: getMyObjectsWithDownstream
      parameters:
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataEntityRefList'
      tags:
        - dataEntity
  /api/dataentities/popular:
    get:
      summary: Get popular entities
      description: Returns list of the most popular data entities throughout the catalog
      operationId: getPopular
      parameters:
        - $ref: './components.yaml/#/components/parameters/PageParam'
        - $ref: './components.yaml/#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataEntityRefList'
      tags:
        - dataEntity
...
  /api/dataentities/{data_entity_id}:
    get:
      summary: Get DataEntity details
      description: "Gets DataEntity's detailed information such as metadata, description, etc"
      operationId: getDataEntityDetails
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataEntityDetails'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/description:
    put:
      summary: Upsert DataEntity's internal description
      description: Upserts DataEntity's internal description in markdown format
      operationId: upsertDataEntityInternalDescription
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/InternalDescriptionFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/InternalDescription'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/name:
    put:
      summary: Upsert DataEntity's internal name
      description: Upserts DataEntity's internal name
      operationId: upsertDataEntityInternalName
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/InternalNameFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/InternalName'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/ownership:
    post:
      summary: Create new ownership
      description: Creates new ownership
      operationId: createOwnership
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/OwnershipFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/Ownership'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/ownership/{ownership_id}:
    put:
      summary: Update existing ownership
      description: Updates existing ownership
      operationId: updateOwnership
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
        - $ref: './components.yaml/#/components/parameters/DataEntityOwnershipIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/OwnershipUpdateFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/Ownership'
      tags:
        - dataEntity
    delete:
      summary: Delete ownership
      description: Deletes the ownership
      operationId: deleteOwnership
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
        - $ref: './components.yaml/#/components/parameters/DataEntityOwnershipIdParam'
      responses:
        '204':
          $ref: './components.yaml/#/components/responses/Deleted'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/tags:
    put:
      summary: Creates tags relations for DataEntity entity
      description: Creates tags relations tags for DataEntity entity. Also creates corresponding tags in the system if they don't exist.
      operationId: createDataEntityTagsRelations
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/DataEntityTagsFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/TagList'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/metadata:
    post:
      summary: Create DataEntity's metadata field value
      description: Creates DataEntity's metadata field value
      operationId: createDataEntityMetadataFieldValue
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/MetadataFieldValueFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/MetadataFieldValueList'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/metadata/{metadata_field_id}:
    put:
      summary: Upsert DataEntity's metadata field value
      description: Upserts DataEntity's metadata field value
      operationId: upsertDataEntityMetadataFieldValue
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
        - $ref: './components.yaml/#/components/parameters/DataEntityMetadataFieldIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components.yaml/#/components/schemas/MetadataFieldValueUpdateFormData'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/MetadataFieldValue'
      tags:
        - dataEntity
    delete:
      summary: "Delete DataEntity's metadata field value"
      description: "Delete DataEntity's metadata field value"
      operationId: deleteDataEntityMetadataFieldValue
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
        - $ref: './components.yaml/#/components/parameters/DataEntityMetadataFieldIdParam'
      responses:
        '204':
          $ref: './components.yaml/#/components/responses/Deleted'
      tags:
        - dataEntity
  /api/dataentities/{data_entity_id}/lineage:
    get:
      summary: Get DataEntity's lineage graph
      description: Gets DataEntity's lineage graph which size depends on lineage_depth parameter
      operationId: getDataEntityLineage
      parameters:
        - $ref: './components.yaml/#/components/parameters/DataEntityIdParam'
        - name: lineage_depth
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: './components.yaml/#/components/schemas/DataEntityLineage'
      tags:
        - dataEntity
...